Conflicts:
	Project/Assembly-CSharp-vs.csproj
	Project/Assembly-CSharp.csproj
	Project/Assets/Scenes/LevelSelectScene.unity
	Project/Library/CurrentMaximizeLayout.dwlt
	Project/Library/guidmapper
	Project/Library/metadata/07/07782307c3b3d414f956e8a948cf6df4
	Project/Library/metadata/3d/3d6764dccac9ec64c9c24f0618d649fb
	Project/Library/metadata/40/40a84c745821c924dafc7a09efc6ecd9

+++ Project/Assets/Scripts/HumanUnitScripts/Jet.cs
@@ -166,7 +166,7 @@ public class Jet : Player {
 		float pauseHeight = Screen.height / 3;
 
 		GUI.Box(new Rect (0, pauseHeight, buttonWidth, Screen.height-pauseHeight),
-		        "GAME INFO\n"+"Charactor: "+roleName()+"\nHP: "+this.HP+"\nBase Damage: "+this.baseDamage+"\nDefence: "+this.baseDefense
+		        "GAME INFO\n"+"Charactor: "+roleName()+"\nHP: "+(int)this.HP+"\nBase Damage: "+this.baseDamage+"\nDefence: "+this.baseDefense
 		        +"\nAttackHitRate: "+this.attackHitRate+"\nMovement Range: "+this.movementRange+"\nAttack Range: "+this.attackRange,"Box");
